requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
wheel
django-allauth==0.43.0
django-rest-auth
django-rest-auth[with_social]
python-decouple